Increase model size towards resource limits of the RFSoC2x2
Note: This is of course the limit with repsect to our folding
configuration, i.e., resource settings and the T^2 cycles per sequence
target for setting the parallelism.

As the number of heads, dimensions and the sequence length is of more
interest for now, the bit-width has been reduced to still fit the model
onto the device.